refactor: split vec_types module
